refactor(step5): 원시값 포장 클래스를 Car 내부에서 외부로 옮김
- Car 내부에서 사용하던 Location, Name
