+++ src/main/java/com/kkambi/racing/domain/Car.java
@@ -1,7 +1,5 @@
 package com.kkambi.racing.domain;
 
-import java.util.Objects;
-
 public class Car {
 
     private static final int criteriaToMove = 4;
@@ -26,57 +24,5 @@ public class Car {
 
     public Name getName() {
         return name;
-    }
-
-    public static class Location implements Comparable<Location> {
-        private final int value;
-
-        public Location(final int value) {
-            if (value < 0) {
-                throw new IllegalArgumentException(Error.INVALID_ARGUMENT.getMessage());
-            }
-            this.value = value;
-        }
-
-        public int getValue() {
-            return value;
-        }
-
-        public Location move() {
-            return new Location(value + 1);
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            Location location = (Location) o;
-            return value == location.value;
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(value);
-        }
-
-        @Override
-        public int compareTo(Location o) {
-            return value - o.getValue();
-        }
-    }
-
-    public static class Name {
-        private final String value;
-
-        public Name(final String value) {
-            if (value.length() > 5) {
-                throw new IllegalArgumentException(Error.INVALID_ARGUMENT.getMessage());
-            }
-            this.value = value;
-        }
-
-        public String getValue() {
-            return value;
-        }
     }
 }
